##2 백엔드 Service, ServiceImpl 분리
