+++ localhost/src/main/java/bssm/spring/board/service/BoardServiceImpl.java
@@ -1,0 +1,88 @@
+package bssm.spring.board.service;
+
+import bssm.spring.board.domain.Board;
+import bssm.spring.board.dto.BoardDto;
+import bssm.spring.board.repository.BoardRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class BoardServiceImpl implements BoardService{
+
+    private final BoardRepository boardRepository;
+
+    @Transactional
+    @Override
+    public void savePost(BoardDto boardDto){
+        boardRepository.save(boardDto.ToEntity());
+    }
+
+    @Transactional
+    @Override
+    public List<BoardDto> getBoardList(){
+        List<Board> all = boardRepository.findAll();
+        List<BoardDto> boardDtoList = new ArrayList<>();
+
+        for(Board board : all){
+            BoardDto boardDto = BoardDto.builder()
+                    .id(board.getId())
+                    .title(board.getTitle())
+                    .content(board.getContent())
+                    .writer(board.getWriter())
+                    .localDateTime(board.getLocalDateTime())
+                    .build();
+
+            boardDtoList.add(boardDto);
+        }
+
+        return boardDtoList;
+    }
+
+    @Transactional
+    @Override
+    public BoardDto getPost(Long id){
+        Optional<Board> boardWrapper = boardRepository.findById(id);
+        Board board = boardWrapper.get();
+
+        return BoardDto.builder()
+                .id(board.getId())
+                .title(board.getTitle())
+                .content(board.getContent())
+                .writer(board.getWriter())
+                .localDateTime(board.getLocalDateTime())
+                .build();
+    }
+
+    @Transactional
+    @Override
+    public void deletePost(Long id){
+        boardRepository.deleteById(id);
+    }
+
+    @Transactional
+    @Override
+    public List<BoardDto> searchPosts(String keyword){
+        List<Board> boards = boardRepository.findByTitleContaining(keyword);
+        List<BoardDto> boardList = new ArrayList<>();
+
+        for(Board board : boards){
+            BoardDto build = BoardDto.builder()
+                    .id(board.getId())
+                    .title(board.getTitle())
+                    .content(board.getContent())
+                    .writer(board.getWriter())
+                    .localDateTime(board.getLocalDateTime())
+                    .build();
+
+            boardList.add(build);
+        }
+
+        return boardList;
+    }
+}
